docs: update readme

+++ run.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
-rm -r ./target
+if [ -d "./target" ]
+then
+  rm -r ./target
+fi
+
 docker run --rm -v $(pwd)/contracts:/home/app/contracts cosmonaut:1.0.0 $1
